requirements.txt
moviepy>=1.0.0
Pillow>=8.0.0
python-telegram-bot>=13.0
PyYAML>=5.0